fixed dependencies for the shared psql plugin in package qt
The postgresql library -lpq depends on -lsecure32, -lssl, and so
on. These are static libraries. When we build the psql plugin as
a shared library, linking fails unless we provide the list of
required libraries.

Previously, this was taken care of with a hack in the configure
script. Since that hack is not suitable for an upstream merge
request, I have moved it here.