fix use of pthreads in package vmime
The configure option "-without-pthread" seems not to work even
if it is listed as a configure option. As long as pthreads is
available, it is detected. This meant that the configuration
of vmime dependend on whether pthreads had already been installed,
even though pthreads was not listed as a dependency.

The solution seems to be to add the missing #include <pthread.h>
and allow pthreads to be used.

Also now build one of the examples as test-vmime.exe.