LLVM: Update to 18.1.4
* src/llvm.mk: Update version and checksum. It is no longer possible to build
a shared libLLVM.dll for Windows in newer versions of LLVM because the maximum
number of symbols that can be exported from a DLL is exceeded. Build static
libraries instead. Set a windows-gnu target because the library naming scheme
that llvm-config expects for windows-msvc (the default) is different.
No changes for Linux build rule.

* src/mesa.mk: Use static LLVM libraries when building for Windows.